Refactored the scoring system to use floating-point numbers and updated the database schema.
- **Database Changes**:
    - Incremented the database version to `6` in `AppDatabase.kt`.
    - Enabled `fallbackToDestructiveMigration` in the Android database configuration.
    - Updated the `Scores` entity to use `Float` for the `score` field instead of `Int`.

- **Score Management & Logic**:
    - Updated `ScoresViewModel` and various game components (`Generala`, `Truco`, `Levels`, `Ranking`, `Points`) to handle `Float` scores.
    - Implemented a `formatScore` helper function in `Points.kt` to display abbreviated values (e.g., "1k" for 1,000, "1.5M" for 1,500,000) and strip unnecessary decimals.
    - Refined total score calculations across different game modes to support double-precision summation before converting back to float or int for display.

- **UI Improvements**:
    - **Generala**: Improved score rendering logic to handle null/placeholder values and updated color-coding for specific score types.
    - **Levels**: Added a `Spacer` to the layout and updated level increment/decrement logic to maintain floating-point consistency.
    - **Points**: Updated score entry and editing dialogs to accept floating-point input.

+++ GamesScoringApp/composeApp/src/commonMain/kotlin/com/example/games_scoring_app/Data/Scores.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 data class Scores(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
     @ColumnInfo(name = "id_player") val id_player: Int,
-    @ColumnInfo(name = "score") var score: Int,
+    @ColumnInfo(name = "score") var score: Float,
     @ColumnInfo(name = "isTotalScore") val isFinalScore: Boolean,
     @ColumnInfo(name = "id_score_type") val id_score_type: Int
 )
